python/invoker_client/requirements.txt
grpcio-tools==1.56.0